Changing Rule properties to a Properties object
Add the AddPropertyBag to Rule
fix some minor issues with the AddPropertie on run
